# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BooksView": {
        "properties": {
          "baseline": {
            "title": "Baseline",
            "type": "string"
          },
          "books": {
            "items": {
              "$ref": "#/components/schemas/BookComparison"
            },
            "title": "Books",
            "type": "array"
          },
          "experiment": {
            "title": "Experiment",
            "type": "string"
          }
        },
        "required": [
          "experiment",
          "baseline",
          "books"
        ],
        "title": "BooksView",
        "type": "object"
      },
      "CalibrationView": {
        "properties": {
          "fitted": {
            "title": "Fitted",
            "type": "boolean"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "temperatures": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Temperatures",
            "type": "object"
          }
        },
        "required": [
          "fitted",
          "temperatures",
          "meta"
        ],
        "title": "CalibrationView",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Every error response: generic, never echoing input or exception text.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "required": [
          "error",
          "fields"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "RoleModels": {
        "properties": {
          "chain": {
            "items": {
              "type": "string"
            },
            "title": "Chain",
            "type": "array"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelHealth"
            },
            "title": "Models",
            "type": "array"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "enabled",
          "chain",
          "models"
        ],
        "title": "RoleModels",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RakshaQuant Web Console",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai/calibration": {
      "get": {
        "operationId": "ai_calibration_api_ai_calibration_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or wrong token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Ai Calibration"
      }
    },
    "/api/ai/calls": {
      "get": {
        "operationId": "ai_calls_api_ai_calls_get",
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z_]{1,32}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "book",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9]{1,16}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LLMCallRow"
                  },
                  "title": "Response Ai Calls Api Ai Calls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or wrong token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Ai Calls"
      }
    },
    "/api/ai/decision-models": {
      "get": {
        "operationId": "ai_decision_models_api_ai_decision_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionModelStats"
                  },
                  "title": "Response Ai Decision Models Api Ai Decision Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or wrong token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Ai Decision Models"
      }
    },
    "/api/ai/models": {
      "get": {
        "operationId": "ai_models_api_ai_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleModels"
                  },
                  "title": "Response Ai Models Api Ai Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or wrong token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Ai Models"
      }
    },
...
    "/api/books": {
      "get": {
        "operationId": "books_api_books_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or wrong token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Books"
      }
    },
...
  }
}